refactor(spaces): lift vector-index management into spaces/vector-index.ts (A17.7 step 1)
The audit sequences this file deliberately: lift the cohesive vector-index block
first, then the risky lifecycle/rename half. This is step 1.

spaces/vector-index.ts now owns:
  - ensureVectorSearchIndex   — build/diff each collection's Atlas $vectorSearch index
  - pollVectorIndexReady, waitForSpaceIndexesReady, finalizeSpaceIndexReady
                              — poll to READY; finalize flips a space's indexStatus
                                to ready/failed
  - vectorFilterFieldsFor, deriveVectorFilterFields
                              — the filter fields that let a recall use native ANN
                                pre-filtering instead of an exhaustive ENN scan

Dependency is one-way: spaces.ts calls in, vector-index.ts never reaches back
(verified — the only `createSpace` mention in the moved block was a comment).
brain/recall.ts imports vectorFilterFieldsFor from the new module directly.
spaces.ts: 1204 -> 959 lines.

Removed three pieces of pre-existing dead code, each confirmed dead in the
original before touching it: an `asUpdate` import, and an unused
`const embCfg = getEmbeddingConfig()` in initSpace — the file declared embCfg
twice and only the one inside the vector-index code was ever read. The repo build
has noUnusedLocals off, so none of it ever surfaced.

NOT in this change: the lifecycle (initSpace/createSpace/removeSpace/wipeSpace/
reconcilePendingSpaceOp) and rename (slugify/moveSpaceData/renameSpace) splits.
That half mutates live Config and coordinates with PendingSpaceOp crash recovery
— it deserves its own PR. Tracked as A17.7 step 2.

Verified: build clean, no dead code (tsc --noUnusedLocals); completeness assertion
confirms every non-import source line survived the move. standalone 801/0 (805),
sync 191/0, red-team 272/272, brain 183/183, mcp-tools 94/94, proxy-spaces 44/44,
space-export 14/14, schema 29+85, conflict 15, conversion 8. Docs checked — no doc
references these internals, and the documented indexStatus contract
(building -> ready/failed) is unchanged.
